// src/app/core/interfaces/passengers.ts
export interface clients {
  id:number,
    name: string;             
    phone: string;            
    photo: string;             
    ticketNumber: string;      
    cinema: string;            
    noOfPersons: number;       
    price: number;              
    noOfSelling: string;       
    ticket: string;            
  }